requirements.txt
python-telegram-bot
ffmpegio
python-dotenv